refactor: Extract header into separate component and remove test code
🏗️ COMPONENT ARCHITECTURE IMPROVEMENT

✨ Header Component Extraction:
- Create dedicated Header component (client/src/components/ui/Header.tsx)
- Move all header logic from App.tsx to Header component
- Include signer wallet management functionality
- Maintain network selector and all existing features

🧹 Code Organization:
- Remove header-related state and functions from App.tsx
- Clean separation of concerns between App and Header
- Proper component props interface for Header
- Maintain all existing functionality with better structure

🗑️ Test Code Cleanup:
- Remove client/src/utils/testTokenDetection.ts file
- Remove testTokenDetection imports from WalletPage.tsx
- Remove test function window assignments
- Clean up unused Badge import from App.tsx

🎯 Benefits:
- Better component organization and maintainability
- Cleaner App.tsx with focused responsibilities
- Reusable Header component for future use
- Removed development-only test code
- Improved code structure and readability

Technical Implementation:
- Header component manages its own state for signer functionality
- Props interface for communication with parent App component
- Proper TypeScript types and error handling
- All existing features preserved in new structure

This refactoring improves code organization while maintaining
all existing functionality and removing unnecessary test code.
